Yield-Per-Recruit

Yield-Per-Recruit

ypr(tmax = 20L, k = 0.15, Linf = 100, t0 = 0, a = 1e-05, b = 3,
  Lm = Linf/2, fa = 1, fb = 1, Rt = 1L, nu = 0.2, Lv = Linf/2,
  Llo = 0, Lup = Linf, mu = 0.1, rho = 0, eta = 0.1)

Arguments

tmax
The maximum age.

k
The growth coefficient.

Linf
The mean maximum length.

t0
The age at length 0.

a
The condition factor.

b
The scaling exponent.

Lm
The length at maturity.

fa
The fecundity factor.

fb
The fecundity scaling exponent.

Rt
A count of the recruitment age for the stock recruitment relationship.

nu
The annual probability of dying due to natural causes.

Lv
The length at which vulnerable to harvest.

Llo
The lower harvest length.

Lup
The upper harvest length.

mu
The annual probability of being captured.

rho
The release probability.

eta
The hooking mortality probability.

Examples

ypr()

#> # A tibble: 20 x 10
#>      Age Length Weight Fecundity Mortality Survivorship Vulnerable Harvest
#>    <int>  <dbl>  <dbl>     <dbl>     <dbl>        <dbl> <lgl>      <lgl>  
#>  1     1   13.9 0.0270      0          0.2       1      FALSE      FALSE  
#>  2     2   25.9 0.174       0          0.2       0.8    FALSE      FALSE  
#>  3     3   36.2 0.476       0          0.2       0.64   FALSE      FALSE  
#>  4     4   45.1 0.918       0          0.2       0.512  FALSE      FALSE  
#>  5     5   52.8 1.47        1.47       0.2       0.410  TRUE       TRUE   
#>  6     6   59.3 2.09        2.09       0.2       0.328  TRUE       TRUE   
#>  7     7   65.0 2.75        2.75       0.2       0.262  TRUE       TRUE   
#>  8     8   69.9 3.41        3.41       0.2       0.210  TRUE       TRUE   
#>  9     9   74.1 4.06        4.06       0.2       0.168  TRUE       TRUE   
#> 10    10   77.7 4.69        4.69       0.2       0.134  TRUE       TRUE   
#> 11    11   80.8 5.27        5.27       0.2       0.107  TRUE       TRUE   
#> 12    12   83.5 5.82        5.82       0.2       0.0859 TRUE       TRUE   
#> 13    13   85.8 6.31        6.31       0.2       0.0687 TRUE       TRUE   
#> 14    14   87.8 6.76        6.76       0.2       0.0550 TRUE       TRUE   
#> 15    15   89.5 7.16        7.16       0.2       0.0440 TRUE       TRUE   
#> 16    16   90.9 7.52        7.52       0.2       0.0352 TRUE       TRUE   
#> 17    17   92.2 7.84        7.84       0.2       0.0281 TRUE       TRUE   
#> 18    18   93.3 8.12        8.12       0.2       0.0225 TRUE       TRUE   
#> 19    19   94.2 8.36        8.36       0.2       0.0180 TRUE       TRUE   
#> 20    20   95.0 8.58        8.58       1         0.0144 TRUE       TRUE   
#> # ... with 2 more variables: FishedMortality <dbl>, FishedSurvivorship <dbl>