Stock-Recruitment Parameters

Stock-Recruitment Parameters

# S3 method for ypr_population
ypr_tabulate_sr(object, Ly = 0,
  harvest = TRUE, biomass = FALSE, all = FALSE, ...)

Arguments

object
The populations or populations.

Ly
The minimum length (trophy) fish to consider when calculating the yield.

harvest
A flag indicating whether to calculate the yield for harvested fish or captures.

biomass
A flag indicating whether to calculate the yield in terms of the biomass versus number of individuals.

all
A flag indicating whether to include all parameter values.

...
Unused parameters.

Value

A data frame of stock-recruitment parameters.

Examples

ypr_tabulate_sr(ypr_population()) # Beverton-Holt

#> # A tibble: 3 x 7
#>   Type        pi     u  Eggs Recruits Spawners Fecundity
#>   <chr>    <dbl> <dbl> <dbl>    <dbl>    <dbl>     <dbl>
#> 1 unfished 0     0     2981.    0.667    1.58      3764.
#> 2 actual   0.2   0.2    231.    0.134    0.175     2650.
#> 3 optimal  0.106 0.106 1075.    0.419    0.697     3085.

ypr_tabulate_sr(ypr_population(BH = 0L)) # Ricker

#> # A tibble: 3 x 7
#>   Type        pi     u  Eggs Recruits Spawners Fecundity
#>   <chr>    <dbl> <dbl> <dbl>    <dbl>    <dbl>     <dbl>
#> 1 unfished 0     0     4451.    0.995    2.36      3764.
#> 2 actual   0.2   0.2    585.    0.340    0.441     2650.
#> 3 optimal  0.124 0.124 1869.    0.791    1.25      2991.