Yield

Calculates the yield for a population.

Usage

ypr_yield(object, Ly = 0, harvest = TRUE, biomass = FALSE, ...)

Arguments

object

The population or populations.

Ly

The minimum length (trophy) fish to consider when calculating the yield (cm).

harvest

A flag specifying whether to calculate the yield for harvested fish or captures.

biomass

A flag specifying whether to calculate the yield in terms of the biomass versus number of individuals.

...

Unused parameters.

Value

The yield as number of fish or biomass.

Details

By default, with Rmax = 1 the number of individuals is the proportion of the recruits at the carrying capacity. If the yield is given in terms of the biomass (kg) then the scaling also depends on the value of Wa (g).

See also

Examples

ypr_yield(ypr_population())
#> [1] 0.04220691
#> attr(,"Age")
#> [1] 6.7687
#> attr(,"Length")
#> [1] 62.13068
#> attr(,"Weight")
#> [1] 2580.272
#> attr(,"Effort")
#> [1] 2.117905
ypr_yield(ypr_ecotypes(Linf = c(100, 200)))
#> [1] 0
#> attr(,"Age")
#> [1] NA
#> attr(,"Length")
#> [1] NA
#> attr(,"Weight")
#> [1] NA
#> attr(,"Effort")
#> [1] 2.117905