Chilliwack Lake Bull Trout Populations Parameters (2005)

The populations parameters for Bull Trout in Chilliwack Lake from Taylor (2005)

chilliwack_bt_05

Format

An object of class ypr_populations().

References

Taylor, J.L. 2005. Sustainability of the Chilliwack Lake Char Fishery. Ministry of Water, Land and Air Protection, Surrey, B.C.

Examples

chilliwack_bt_05

#> tmax:  20
#> k:     0.12
#> Linf:  86
#> t0:    0.018
#> k2:    0.12
#> Linf2: 86
#> L2:    1000
#> Wb:    2.8
#> Ls:    38.5
#> Sp:    100
#> es:    1
#> Sm:    0
#> fb:    1
#> tR:    1
#> BH:    1
#> Rk:    2.5, 4.6, 2.5, 4.6, 2.5, 4.6, 2.5, 4.6, 2.5, 4.6, 2.5, 4.6
#> n:     0.17
#> nL:    0.17
#> Ln:    50
#> Lv:    36.5
#> Vp:    11.7
#> Llo:   0, 0, 60, 60, 1000, 1000, 0, 0, 60, 60, 1000, 1000
#> Lup:   1000
#> Nc:    0.15
#> pi:    0.2
#> rho:   0
#> Hm:    0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.2, 0.2, 0.2, 0.2, 0.2, 0.2
#> Rmax:  250
#> Wa:    0.022
#> fa:    2.29
#> q:     0.1

yield <- ypr_tabulate_yield(chilliwack_bt_05, type = "optimal")
yield$pi <- round(yield$pi, 2)
yield <- yield[c("Llo", "Hm", "Rk", "pi")]
yield <- tidyr::spread(yield, Rk, pi)
yield <- yield[order(-yield$Hm), ]
yield

#> # A tibble: 6 x 4
#>     Llo    Hm `2.5` `4.6`
#>   <dbl> <dbl> <dbl> <dbl>
#> 1     0  0.2   0.08  0.14
#> 2    60  0.2   0.14  0.25
#> 3  1000  0.2   0.25  0.45
#> 4     0  0.05  0.08  0.14
#> 5    60  0.05  0.17  0.37
#> 6  1000  0.05  0.41  0.74

if (FALSE) {
ypr_plot_yield(chilliwack_bt_05, plot_values = FALSE) +
  ggplot2::facet_grid(Rk ~ Hm) +
  ggplot2::aes(group = Llo, linetype = Llo)
}