Parameter Descriptions for chk Functions

Parameter Descriptions for chk Functions

Arguments

tmax
The maximum age (yr).

k
The VB growth coefficient (per yr).

Linf
The VB mean maximum length (cm).

t0
The (theoretical) age at zero length (yr).

Wb
The weight (as a function of length) scaling exponent.

Ls
The length at which 50% mature (cm).

Sp
The maturity (as a function of length) power.

es
The annual probability of a mature fish spawning.

Sm
The spawning mortality probability.

fb
The fecundity (as a function of weight) scaling exponent.

tR
The age from which survival is density-independent (yr).

BH
Recruitment follows a Beverton-Holt (1) or Ricker (0) relationship.

Rk
The lifetime spawners per spawner at low density.

M
The instantaneous mortality rate (per yr).

Mb
The instantaneous mortality rate (as a function of length) scaling exponent.

Lv
The length at which 50% vulnerable to harvest (cm).

Vp
The vulnerability to harvest (as a function of length) power.

Llo
The lower harvest slot length (cm).

Lup
The upper harvest slot length (cm).

Nc
The slot limits non-compliance probability.

pi
The annual capture probability.

rho
The release probability.

Hm
The hooking mortality probability.

Rmax
The number of recruits at the carrying capacity (ind).

A0
The initial post age tR density independent mortality probability.

Wa
The (extrapolated) weight of a 1 cm individual (g).

fa
The (theoretical) fecundity of a 1 g female (eggs).

q
The catchability (annual probability of capture) for a unit of effort.

all
A flag indicating whether to include all parameter values.

u
A flag indicating whether to plot the exploitation rate as opposed to the capture rate.

percent
A flag specifying whether to plot the number of fish as a percent or frequency (the default).

color
A string of the color around each bar (or NULL).

population
An object of class ypr_population().

plot_values
A flag indicating whether to plot the actual and optimal values.

Ly
The minimum length (trophy) fish to consider when calculating the yield.

harvest
A flag indicating whether to calculate the yield for harvested fish or captures.

biomass
A flag indicating whether to calculate the yield in terms of the biomass versus number of individuals.

title
A string of the report title.

date
A date of the report date.

file
A string of the file name.

binwidth
A positive integer of the width of the bins for grouping.

type
A string indicating whether to include 'both' or just the 'actual' or 'optimal' yield.

object
The populations or populations.

...
Unused parameters.