Population Parameters

Generates an object of class ypr_population.

ypr_population(tmax = 20L, k = 0.15, Linf = 100, t0 = 0, Wb = 3,
  Ls = Linf/2, Sp = 100, es = 1, Sm = 0, fb = 1, tR = 1L,
  BH = 1L, Rk = 3, M = 0.2, Mb = 0, Lv = Linf/2, Vp = 100,
  Llo = 0, Lup = Linf, Nc = 0, pi = 0.2, rho = 0, Hm = 0,
  Rmax = 1, Wa = 0.01, fa = 1, Ea = 1, Eb = 1)

Arguments

tmax
The maximum age (yr).

k
The VB growth coefficient (per yr).

Linf
The VB mean maximum length (cm).

t0
The (theoretical) age at zero length (yr).

Wb
The weight (as a function of length) scaling exponent.

Ls
The length at which 50% mature (cm).

Sp
The maturity (as a function of length) power.

es
The annual probability of a mature fish spawning.

Sm
The spawning mortality probability.

fb
The fecundity (as a function of weight) scaling exponent.

tR
The age from which survival is density-independent (yr).

BH
Recruitment follows a Beverton-Holt (1) or Ricker (0) relationship.

Rk
The numbers of spawners per spawner at low density.

M
The instantaneous mortality rate (per yr).

Mb
The instantaneous mortality rate (as a function of length) scaling exponent.

Lv
The length at which 50% vulnerable to harvest (cm).

Vp
The vulnerability to harvest (as a function of length) power.

Llo
The lower harvest slot length (cm).

Lup
The upper harvest slot length (cm).

Nc
The slot limits non-compliance probability.

pi
The annual capture probability.

rho
The release probability.

Hm
The hooking mortality probability.

Rmax
The number of recruits at the carrying capacity (ind).

Wa
The (extrapolated) weight of a 1 cm individual (g).

fa
The (theoretical) fecundity of a 1 g female (eggs).

Ea
The effort at a pi of 0.01.

Eb
The effort (as a function of pi) scaling exponent.

Value

An object of class ypr_population.

See also

Examples

ypr_population(k = 0.1, Linf = 90)

#> tmax: 20
#> k:    0.1
#> Linf: 90
#> t0:   0
#> Wb:   3
#> Ls:   45
#> Sp:   100
#> es:   1
#> Sm:   0
#> fb:   1
#> tR:   1
#> BH:   1
#> Rk:   3
#> M:    0.2
#> Mb:   0
#> Lv:   45
#> Vp:   100
#> Llo:  0
#> Lup:  90
#> Nc:   0
#> pi:   0.2
#> rho:  0
#> Hm:   0
#> Rmax: 1
#> Wa:   0.01
#> fa:   1
#> Ea:   1
#> Eb:   1